+++ plugin-dev/Source/Sentry/Private/Android/Java/SentryBridgeJava.java
@@ -218,6 +218,10 @@ public class SentryBridgeJava {
 		event.getContexts().remove(key);
 	}
 
+	public static Object getScopeContext(final IScope scope, final String key) {
+		return scope.getContexts().get(key);
+	}
+
 	public static void setScopeExtra(final IScope scope, final String key, final Object values) {
 		scope.getExtras().put(key, values);
 	}