+++ plugin-dev/Source/Sentry/Private/GenericPlatform/GenericPlatformSentryScope.h
@@ -32,7 +32,9 @@ public:
 	virtual TArray<FString> GetFingerprint() const override;
 	virtual void SetLevel(ESentryLevel level) override;
 	virtual ESentryLevel GetLevel() const override;
-	virtual void SetContext(const FString& key, const TMap<FString, FString>& values) override;
+	virtual void SetContext(const FString& key, const TMap<FString, FSentryVariant>& values) override;
+	virtual TMap<FString, FSentryVariant> GetContext(const FString& key) const override;
+	virtual bool TryGetContext(const FString& key, TMap<FString, FSentryVariant>& value) const override;
 	virtual void RemoveContext(const FString& key) override;
 	virtual void SetExtra(const FString& key, const FSentryVariant& value) override;
 	virtual FSentryVariant GetExtra(const FString& key) const override;
@@ -53,7 +55,7 @@ private:
 	TMap<FString, FString> Tags;
 	TMap<FString, FSentryVariant> Extra;
 
-	TMap<FString, TMap<FString, FString>> Contexts;
+	TMap<FString, TMap<FString, FSentryVariant>> Contexts;
 
 	TRingBuffer<TSharedPtr<FGenericPlatformSentryBreadcrumb>> Breadcrumbs;