+++ plugin-dev/Source/Sentry/Private/Tests/SentryScope.spec.cpp
@@ -15,8 +15,6 @@
 #include "HAL/PlatformSentryEvent.h"
 #include "HAL/PlatformSentryScope.h"
 
-#include "GenericPlatform/Infrastructure/GenericPlatformSentryConverters.h"
-
 TDelegate<void(USentryEvent*)> UScopeTestBeforeSendHandler::OnScopeTestBeforeSendHandler;
 
 #if WITH_AUTOMATION_TESTS
@@ -27,7 +25,7 @@ BEGIN_DEFINE_SPEC(SentryScopeSpec, "Sentry.SentryScope", EAutomationTestFlags::P
 	FString TestEnvironment;
 	TMap<FString, FString> TestTags;
 	TMap<FString, FSentryVariant> TestExtras;
-	TMap<FString, FString> TestContext;
+	TMap<FString, FSentryVariant> TestContext;
 	TArray<FString> TestFingerprint;
 END_DEFINE_SPEC(SentryScopeSpec)
 
@@ -76,6 +74,26 @@ void SentryScopeSpec::Define()
 			SentryScope->RemoveTag(TEXT("Key1"));
 
 			TestEqual("Tag removed", SentryScope->GetTag(TEXT("Key1")), TEXT(""));
+		});
+	});
+
+	Describe("Scope contexts", [this]()
+	{
+		It("should persist their values", [this]()
+		{
+			TMap<FString, FSentryVariant> NonExistingContext;
+			TestFalse("No context with given key available before it was added", SentryScope->TryGetContext(TEXT("TestContext"), NonExistingContext));
+
+			SentryScope->SetContext(TEXT("TestContext"), TestContext);
+
+			TMap<FString, FSentryVariant> RetrievedContext = SentryScope->GetContext(TEXT("TestContext"));
+
+			TestEqual("Scope context retains all values", RetrievedContext.Num(), 2);
+			TestEqual("Scope context 1", RetrievedContext[TEXT("ContextKey1")], TestContext[TEXT("ContextKey1")]);
+			TestEqual("Scope context 2", RetrievedContext[TEXT("ContextKey2")], TestContext[TEXT("ContextKey2")]);
+
+			SentryScope->RemoveContext(TEXT("TestContext"));
+			TestFalse("No context with given key available after it was removed", SentryScope->TryGetContext(TEXT("TestContext"), NonExistingContext));
 		});
 	});
 
@@ -135,7 +153,6 @@ void SentryScopeSpec::Define()
 		});
 	});
 
-#if (PLATFORM_MICROSOFT || PLATFORM_LINUX) && USE_SENTRY_NATIVE
 	Describe("Scope params", [this]()
 	{
 		It("should be applied to event", [this]()
@@ -148,26 +165,14 @@ void SentryScopeSpec::Define()
 
 			UScopeTestBeforeSendHandler::OnScopeTestBeforeSendHandler.BindLambda([this](USentryEvent* SentryEvent)
 			{
-				TSharedPtr<FGenericPlatformSentryEvent> Event = StaticCastSharedPtr<FGenericPlatformSentryEvent>(SentryEvent->GetNativeObject());
-
-				sentry_value_t NativeEvent = Event->GetNativeObject();
-
-				sentry_value_t level = sentry_value_get_by_key(NativeEvent, "level");
-				sentry_value_t fingerprint = sentry_value_get_by_key(NativeEvent, "fingerprint");
-				sentry_value_t tags = sentry_value_get_by_key(NativeEvent, "tags");
-				sentry_value_t extra = sentry_value_get_by_key(NativeEvent, "extra");
-				sentry_value_t contexts = sentry_value_get_by_key(NativeEvent, "contexts");
-
-				sentry_value_t testContext = sentry_value_get_by_key(contexts, "TestContext");
-
-				TestEqual("Event level", FGenericPlatformSentryConverters::SentryLevelToUnreal(level), ESentryLevel::Fatal);
-				TestEqual("Event fingerprint", FGenericPlatformSentryConverters::StringArrayToUnreal(fingerprint), TestFingerprint);
-				TestEqual("Event tags 1", FGenericPlatformSentryConverters::StringMapToUnreal(tags)[TEXT("TagsKey1")], TestTags[TEXT("TagsKey1")]);
-				TestEqual("Event tags 2", FGenericPlatformSentryConverters::StringMapToUnreal(tags)[TEXT("TagsKey2")], TestTags[TEXT("TagsKey2")]);
-				TestEqual("Event extra 1", FGenericPlatformSentryConverters::StringMapToUnreal(extra)[TEXT("ExtrasKey1")], TestExtras[TEXT("ExtrasKey1")]);
-				TestEqual("Event extra 2", FGenericPlatformSentryConverters::StringMapToUnreal(extra)[TEXT("ExtrasKey2")], TestExtras[TEXT("ExtrasKey2")]);
-				TestEqual("Event context 1", FGenericPlatformSentryConverters::StringMapToUnreal(testContext)[TEXT("ContextKey1")], TestContext[TEXT("ContextKey1")]);
-				TestEqual("Event context 2", FGenericPlatformSentryConverters::StringMapToUnreal(testContext)[TEXT("ContextKey2")], TestContext[TEXT("ContextKey2")]);
+				TestEqual("Event level", SentryEvent->GetLevel(), ESentryLevel::Fatal);
+				TestEqual("Event fingerprint", SentryEvent->GetFingerprint(), TestFingerprint);
+				TestEqual("Event tags 1", SentryEvent->GetTags()[TEXT("TagsKey1")], TestTags[TEXT("TagsKey1")]);
+				TestEqual("Event tags 2", SentryEvent->GetTags()[TEXT("TagsKey2")], TestTags[TEXT("TagsKey2")]);
+				TestEqual("Event extra 1", SentryEvent->GetExtras()[TEXT("ExtrasKey1")], TestExtras[TEXT("ExtrasKey1")]);
+				TestEqual("Event extra 2", SentryEvent->GetExtras()[TEXT("ExtrasKey2")], TestExtras[TEXT("ExtrasKey2")]);
+				TestEqual("Event context 1", SentryEvent->GetContext(TEXT("TestContext"))[TEXT("ContextKey1")], TestContext[TEXT("ContextKey1")]);
+				TestEqual("Event context 2", SentryEvent->GetContext(TEXT("TestContext"))[TEXT("ContextKey2")], TestContext[TEXT("ContextKey2")]);
 			});
 
 			USentryEvent* SentryEvent = USentryEvent::Create(CreateSharedSentryEvent());
@@ -185,7 +190,6 @@ void SentryScopeSpec::Define()
 			SentrySubsystem->Close();
 		});
 	});
-#endif
 }
 
 #endif