+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Breaking Changes
 
-- In `SentryScope` class the "Value" suffix was dropped for tags and extras get/set functions.
+- In `SentryScope` class the "Value" suffix has been removed from the get/set functions for tags and extras.
+- The get/set functions for tags, contexts, and extras in `SentryScope` now accept a variant type instead of string for the value.
 
 ## 1.0.0-beta.3